fix(ci): line-limit gate — decompose oversized PR monitor files
